cart api is updated for coustomer delivery charges

+++ customerandorder/src/main/java/com/jippy/customerandorder/repository/CoOrderPriceBreakupRepository.java
@@ -4,12 +4,12 @@ import com.jippy.customerandorder.entity.CoOrderPriceBreakup;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.Optional;
-
 @Repository
-public interface CoOrderPriceBreakupRepository extends JpaRepository<CoOrderPriceBreakup,Long> {
+public interface CoOrderPriceBreakupRepository
+        extends JpaRepository<CoOrderPriceBreakup, Integer> {
 
-//    Fetch price breakup using orderId
+    /**
+     * Fetch price breakup using order ID.
+     */
     CoOrderPriceBreakup findByOrder_OrderId(String orderId);
-
 }
